# Conflicts:
#	src/main/java/io/mailtrap/client/layers/MailtrapEmailTestingApi.java
#	src/main/java/io/mailtrap/factory/MailtrapClientFactory.java

+++ src/main/java/io/mailtrap/api/TestingEmailsImpl.java
@@ -20,7 +20,7 @@ public class TestingEmailsImpl extends SendApiResource implements TestingEmails 
     }
 
     @Override
-    public SendResponse send(MailtrapMail mail, int inboxId) {
+    public SendResponse send(MailtrapMail mail, long inboxId) {
         validateRequestBodyOrThrowException(mail);
         RequestData requestData = new RequestData();
         if (mail.getHeaders() != null) {
